Refactor code to separate WholeGraph-related functions
