+++ python/graphstorm/dataloading/utils.py
@@ -137,13 +137,3 @@ def flip_node_mask(dist_tensor, indices):
         part_policy=dist_tensor.part_policy)
     flipped_dist_tensor[indices] = 1 - dist_tensor[indices]
     return flipped_dist_tensor
-
-def is_wholegraph_embedding(data):
-    """ Check if the data is in WholeMemory emedding format which
-        is required to use wholegraph framework.
-    """
-    try:
-        import pylibwholegraph
-        return isinstance(data, pylibwholegraph.torch.WholeMemoryEmbedding)
-    except: # pylint: disable=bare-except
-        return False
